Move IPersistProperty bag to Interop
